Merged PR 569: [Feature] Added optional configuration to MuBuild to use git reference/omnicache
- added omnicache to mubuild
- Added omnicache options
- Added fully clone option (Full)

This requires pull request #568 to work properly

You can specify a reference cache for MuBuild as a whole or for specific repos. A good example of how to use the new config option for cloning a specific dependency would be in ConfigValidator
```yaml
Dependencies:
        - Silicon/Arm/Tiano
            Path: Silicon/Arm/Tiano
            Url: https://github.com/Microsoft/mu_silicon_arm_tiano.git
            Branch: release/20180529
            ReferencePath: "../place"
            Full: true
```

Full specifies that the repo should be fully cloned (instead of doing shallow) even if you are using a branch. ReferencePath is how you can specify the omnicache dir (it is relative to the workspace directory). If there is a global reference path, it does not have preference over a dependency reference path (specificity wins).

To use an omnicache in MuBuild globally, you have two options:
```bash
mubuild --omnicache ../dir
mubuild --reference ../dir
```